migration: move utils tests alongside utils.go
splitReplicaDSNs and maskPasswordInDSN were moved from runner.go to
utils.go; move their existing table-driven tests (TestSplitReplicaDSNs,
TestPasswordMasking) from migration_test.go into a new utils_test.go.
TestDSN stays in migration_test.go since it tests Runner.dsn().
